Provisioning for Version 1.1.0 (February 11, 2021) (13:43)
"So close... We've had so many rewrites, changes, complications. We had a version in Python with error codes, then separated handlers. Then we made Golang port, then a rewrite, after that one we had one with a bad functions. Due to an oversight, had to rewrite again. Afterwards I chose to adopt a VBScript-like syntax and broke it twice more (mainly because of the variable type that VBScript uses, variant) before ditching that too. Now we have a version which I think will be great, it has support for multiple variable types, extendable programming language support, bi-directional piping (only one side at a time though), command metadata, string references (this really is more like a limitation turned into a feature, but I'll take what I can get), and commented source code. I think I had a few run-on sentences but that doesn't matter, this is meant to tell the history of Majonez."