Provisioning for Version 1.1.0 (February 11, 2021) (17:21)
"I use this file in a lot of other projects because these are common functions I use. You may notice that there is no errors returned, I don't feel a need to return an error along with its output. I like combining them both at one (with an exception being strings) so I don't have to bother with checking if error is nil"