Provisioning for Version 1.1.0 (February 19, 2021) (22:31)
"2 things of note: the fact I have done everything up to now is kind of sad & the whole provisioning for 1.1.0 is starting to get old. I hope that bug testing is quick and that I can push 1.1.0 so I won't have to put the stupid provisioning thing over and over. I messed it up at one point and got the dates wrong and that was no good."