Version 1.1.0 (February 21, 2021) (22:58)
"I wanted to be sure that I put the history here. I'm not really a fan of changelogs, but I am interested in the development of software. I have made a bunch of mistakes when committing so I am creating a file that can be relied on for the dates of the versions. All the comments on the commits are correct. (At the moment) I will try making sure that the comments are 100% correct. The date and time will be in this file instead."